Added latest error message to Main Flight Display
