+++ src/QmlControls/MavManager.cc
@@ -32,6 +32,7 @@ This file is part of the QGROUNDCONTROL project
 #include "UASManager.h"
 #include "Waypoint.h"
 #include "MavManager.h"
+#include "UASMessageHandler.h"
 
 #define UPDATE_TIMER 50
 #define DEFAULT_LAT  38.965767f
@@ -40,6 +41,12 @@ This file is part of the QGROUNDCONTROL project
 MavManager::MavManager(QObject *parent)
     : QObject(parent)
     , _mav(NULL)
+    , _currentMessageCount(0)
+    , _messageCount(0)
+    , _currentErrorCount(0)
+    , _currentWarningCount(0)
+    , _currentNormalCount(0)
+    , _currentMessageType(MessageNone)
     , _roll(0.0f)
     , _pitch(0.0f)
     , _heading(0.0f)
@@ -99,6 +106,8 @@ QString MavManager::loadSetting(const QString &name, const QString& defaultValue
 void MavManager::_forgetUAS(UASInterface* uas)
 {
     if (_mav != NULL && _mav == uas) {
+        // Stop listening for system messages
+        disconnect(UASMessageHandler::instance(), &UASMessageHandler::textMessageCountChanged,  this, &MavManager::_handleTextMessage);
         // Disconnect any previously connected active MAV
         disconnect(_mav, SIGNAL(attitudeChanged                     (UASInterface*, double,double,double,quint64)),             this, SLOT(_updateAttitude(UASInterface*, double, double, double, quint64)));
         disconnect(_mav, SIGNAL(attitudeChanged                     (UASInterface*, int,double,double,double,quint64)),         this, SLOT(_updateAttitude(UASInterface*,int,double, double, double, quint64)));
@@ -149,6 +158,8 @@ void MavManager::_setActiveUAS(UASInterface* uas)
         emit heartbeatTimeoutChanged();
         // Set new UAS
         _mav = uas;
+        // Listen for system messages
+        connect(UASMessageHandler::instance(), &UASMessageHandler::textMessageCountChanged, this, &MavManager::_handleTextMessage);
         // Now connect the new UAS
         connect(_mav, SIGNAL(attitudeChanged                    (UASInterface*,double,double,double,quint64)),              this, SLOT(_updateAttitude(UASInterface*, double, double, double, quint64)));
         connect(_mav, SIGNAL(attitudeChanged                    (UASInterface*,int,double,double,double,quint64)),          this, SLOT(_updateAttitude(UASInterface*,int,double, double, double, quint64)));
@@ -635,5 +646,89 @@ void MavManager::_waypointViewOnlyListChanged()
             emit latitudeChanged();
         }
         */
+    }
+}
+
+void MavManager::_handleTextMessage(int newCount)
+{
+    // Reset?
+    if(!newCount) {
+        _currentMessageCount = 0;
+        _currentNormalCount  = 0;
+        _currentWarningCount = 0;
+        _currentErrorCount   = 0;
+        _messageCount        = 0;
+        _currentMessageType  = MessageNone;
+        emit newMessageCountChanged();
+        emit messageTypeChanged();
+        emit messageCountChanged();
+        return;
+    }
+
+    UASMessageHandler* pMh = UASMessageHandler::instance();
+    Q_ASSERT(pMh);
+    MessageType_t type = newCount ? _currentMessageType : MessageNone;
+    int errorCount     = _currentErrorCount;
+    int warnCount      = _currentWarningCount;
+    int normalCount    = _currentNormalCount;
+    //-- Add current message counts
+    errorCount  += pMh->getErrorCount();
+    warnCount   += pMh->getWarningCount();
+    normalCount += pMh->getNormalCount();
+    //-- See if we have a higher level
+    if(errorCount != _currentErrorCount) {
+        _currentErrorCount = errorCount;
+        type = MessageError;
+    }
+    if(warnCount != _currentWarningCount) {
+        _currentWarningCount = warnCount;
+        if(_currentMessageType != MessageError) {
+            type = MessageWarning;
+        }
+    }
+    if(normalCount != _currentNormalCount) {
+        _currentNormalCount = normalCount;
+        if(_currentMessageType != MessageError && _currentMessageType != MessageWarning) {
+            type = MessageNormal;
+        }
+    }
+    int count = _currentErrorCount + _currentWarningCount + _currentNormalCount;
+    if(count != _currentMessageCount) {
+        _currentMessageCount = count;
+        // Display current total new messages count
+        emit newMessageCountChanged();
+    }
+    if(type != _currentMessageType) {
+        _currentMessageType = type;
+        // Update message level
+        emit messageTypeChanged();
+    }
+    // Update message count (all messages)
+    if(newCount != _messageCount) {
+        _messageCount = newCount;
+        emit messageCountChanged();
+    }
+    QString errMsg = pMh->getLatestError();
+    if(errMsg != _latestError) {
+        _latestError = errMsg;
+        emit latestErrorChanged();
+    }
+}
+
+void MavManager::resetMessages()
+{
+    // Reset Counts
+    int count = _currentMessageCount;
+    MessageType_t type = _currentMessageType;
+    _currentErrorCount   = 0;
+    _currentWarningCount = 0;
+    _currentNormalCount  = 0;
+    _currentMessageCount = 0;
+    _currentMessageType = MessageNone;
+    if(count != _currentMessageCount) {
+        emit newMessageCountChanged();
+    }
+    if(type != _currentMessageType) {
+        emit messageTypeChanged();
     }
 }
