+++ src/ui/flightdisplay/FlightDisplay.qml
@@ -487,6 +487,15 @@ Item {
         z:                  10
     }
 
+    QGCHudMessage {
+        id:     hudMessage
+        y:      ScreenTools.pixelSizeFactor * (5)
+        width:  (parent.width - 520 > 200) ? parent.width - 520 : 200
+        height: ScreenTools.pixelSizeFactor * (30)
+        anchors.horizontalCenter: parent.horizontalCenter
+        z:      mapBackground.z + 1
+    }
+
     QGCCompassInstrument {
         id:                 compassInstrument
         y:                  ScreenTools.pixelSizeFactor * (5)
@@ -494,7 +503,7 @@ Item {
         size:               ScreenTools.pixelSizeFactor * (160)
         heading:            isNaN(MavManager.heading) ? 0 : MavManager.heading
         visible:            mapBackground.visible && showCompass
-        z:                  mapBackground.z + 1
+        z:                  mapBackground.z + 2
         onResetRequested: {
             y               = ScreenTools.pixelSizeFactor * (5)
             x               = ScreenTools.pixelSizeFactor * (85)
@@ -514,7 +523,7 @@ Item {
         visible:            mapBackground.visible && showAttitudeIndicator
         anchors.right:      root.right
         anchors.rightMargin: ScreenTools.pixelSizeFactor * (85)
-        z:                  mapBackground.z + 1
+        z:                  mapBackground.z + 2
         onResetRequested: {
             y                   = ScreenTools.pixelSizeFactor * (5)
             anchors.right       = root.right
@@ -531,7 +540,6 @@ Item {
         rollAngle:          roll
         pitchAngle:         pitch
         visible:            !mapBackground.visible
-        z:                  10
     }
 
     QGCAttitudeWidget {
