+++ src/ui/toolbar/MainToolBar.h
@@ -47,17 +47,7 @@ class MainToolBar : public QGCQmlWidgetHolder
 {
     Q_OBJECT
     Q_ENUMS(ViewType_t)
-    Q_ENUMS(MessageType_t)
 public:
-    MainToolBar(QWidget* parent = NULL);
-    ~MainToolBar();
-
-    typedef enum {
-        MessageNone,
-        MessageNormal,
-        MessageWarning,
-        MessageError
-    } MessageType_t;
 
     typedef enum {
         ViewNone    = -1,
@@ -66,6 +56,9 @@ public:
         ViewFly    , // MainWindow::VIEW_FLIGHT
         ViewSetup  , // MainWindow::VIEW_SETUP
     } ViewType_t;
+
+    MainToolBar(QWidget* parent = NULL);
+    ~MainToolBar();
 
     Q_INVOKABLE void    onSetupView();
     Q_INVOKABLE void    onPlanView();
@@ -78,9 +71,6 @@ public:
 
     Q_PROPERTY(ViewType_t    currentView        MEMBER _currentView             NOTIFY currentViewChanged)
     Q_PROPERTY(QStringList   configList         MEMBER _linkConfigurations      NOTIFY configListChanged)
-    Q_PROPERTY(MessageType_t messageType        MEMBER _currentMessageType      NOTIFY messageTypeChanged)
-    Q_PROPERTY(int           newMessageCount    MEMBER _currentMessageCount     NOTIFY newMessageCountChanged)
-    Q_PROPERTY(int           messageCount       MEMBER _messageCount            NOTIFY messageCountChanged)
     Q_PROPERTY(int           connectionCount    READ connectionCount            NOTIFY connectionCountChanged)
     Q_PROPERTY(QStringList   connectedList      MEMBER _connectedList           NOTIFY connectedListChanged)
     Q_PROPERTY(bool          showGPS            MEMBER _showGPS                 NOTIFY showGPSChanged)
@@ -104,9 +94,6 @@ signals:
     void connectionCountChanged         (int count);
     void currentViewChanged             ();
     void configListChanged              ();
-    void messageTypeChanged             (MessageType_t type);
-    void newMessageCountChanged         (int count);
-    void messageCountChanged            (int count);
     void connectedListChanged           (QStringList connectedList);
     void showGPSChanged                 (bool value);
     void showMavChanged                 (bool value);
@@ -124,7 +111,6 @@ private slots:
     void _updateConfigurations          ();
     void _linkConnected                 (LinkInterface* link);
     void _linkDisconnected              (LinkInterface* link);
-    void _handleTextMessage             (int newCount);
     void _leaveMessageView              ();
     void _setProgressBarValue           (float value);
     void _updatePixelSize               ();
@@ -141,12 +127,6 @@ private:
     ViewType_t      _currentView;
     QStringList     _linkConfigurations;
     int             _connectionCount;
-    int             _currentMessageCount;
-    int             _messageCount;
-    int             _currentErrorCount;
-    int             _currentWarningCount;
-    int             _currentNormalCount;
-    MessageType_t   _currentMessageType;
     QStringList     _connectedList;
     bool            _showGPS;
     bool            _showMav;
